refactor: move isFormComplete logic to useCardForm and expose via CardContext
