+++ app/src/components/form/CardExpiryDateInput.stories.tsx
@@ -40,6 +40,7 @@ const renderWithContext = () => {
         setCardPassword: () => {},
         cardExpiryDate,
         setCardExpiryDate,
+        isFormComplete: false,
       }}
     >
       <CardExpiryDateInput firstRef={firstRef} onComplete={() => {}} />
