+++ app/src/components/form/CardForm.tsx
@@ -7,26 +7,11 @@ import { CardCVCInput } from './CardCVCInput';
 import { CardPasswordInput } from './CardPasswordInput';
 import { ConfirmButton } from './ConfirmButton';
 import { useCardContext } from '../../hooks/useCardContext';
-import { BrandValidator } from '../../validators/BrandValidator';
 import { useNavigate } from 'react-router-dom';
 import { useFormFocusChain } from '../../hooks/useFormFocusChain';
 
 export function CardForm() {
-  const { cardCompany, cardNumber, cardExpiryDate, cardCVC, cardPassword } = useCardContext();
-  const networkBrand = BrandValidator.detectNetworkBrand(cardNumber.join('')).brand;
-  const lastDigitLength = networkBrand === 'diners' ? 2 : networkBrand === 'amex' ? 3 : 4;
-
-  const isFormComplete =
-    cardCompany !== '' &&
-    cardNumber[0].length === 4 &&
-    cardNumber[1].length === 4 &&
-    cardNumber[2].length === 4 &&
-    cardNumber[3].length === lastDigitLength &&
-    cardExpiryDate['expiry-month'].length === 2 &&
-    cardExpiryDate['expiry-year'].length === 2 &&
-    cardCVC.length === 3 &&
-    cardPassword.length === 2;
-
+  const { cardCompany, cardNumber, isFormComplete } = useCardContext();
   const {
     refs,
     onCardCompanySelected,
