+++ app/src/page/Card.tsx
@@ -11,6 +11,7 @@ export function Card() {
     cardExpiryDate,
     cardCVC,
     cardPassword,
+    isFormComplete,
     setCardCompany,
     setCardNumber,
     setCardExpiryDate,
@@ -26,6 +27,7 @@ export function Card() {
         cardCompany,
         cardCVC,
         cardPassword,
+        isFormComplete,
         setCardNumber,
         setCardExpiryDate,
         setCardCompany,
